+++ src/main/java/org/kdb/inside/brains/view/inspector/model/CanonicalElement.java
@@ -1,0 +1,5 @@
+package org.kdb.inside.brains.view.inspector.model;
+
+public interface CanonicalElement {
+    String getCanonicalName();
+}